# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AnthoLume API v1
  version: 1.0.0
  description: REST API for AnthoLume document management system
servers:
  - url: /api/v1
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    DatabaseInfo:
      type: object
      properties:
        documents_size:
          type: integer
          format: int64
        activity_size:
          type: integer
          format: int64
        progress_size:
          type: integer
          format: int64
        devices_size:
          type: integer
          format: int64
      required:
        - documents_size
        - activity_size
        - progress_size
        - devices_size
...
    BackupType:
      type: string
      enum: [COVERS, DOCUMENTS]
    ImportType:
      type: string
      enum: [DIRECT, COPY]
    OperationType:
      type: string
      enum: [CREATE, UPDATE, DELETE]
    User:
      type: object
      properties:
        id:
          type: string
        admin:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
        - id
        - admin
        - created_at
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
    ImportResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        path:
          type: string
        status:
          type: string
          enum: [FAILED, SUCCESS, EXISTS]
        error:
          type: string
    ImportResultsResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ImportResult'
    DirectoryItem:
      type: object
      properties:
        name:
          type: string
        path:
          type: string
    DirectoryListResponse:
      type: object
      properties:
        current_path:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/DirectoryItem'
    LogEntry:
      type: string
    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/LogEntry'
        filter:
          type: string
        page:
          type: integer
          format: int64
        limit:
          type: integer
          format: int64
        next_page:
          type: integer
          format: int64
        previous_page:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
    InfoResponse:
      type: object
      properties:
        version:
          type: string
        search_enabled:
          type: boolean
        registration_enabled:
          type: boolean
      required:
        - version
        - search_enabled
        - registration_enabled
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: cookie
      name: token
paths:
...
  /info:
    get:
      summary: Get server information
      operationId: getInfo
      tags:
        - Info
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfoResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin:
    get:
      summary: Get admin page data
      operationId: getAdmin
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  database_info:
                    $ref: '#/components/schemas/DatabaseInfo'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Perform admin action (backup, restore, etc.)
      operationId: postAdminAction
      tags:
        - Admin
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [BACKUP, RESTORE, METADATA_MATCH, CACHE_TABLES]
                backup_types:
                  type: array
                  items:
                    $ref: '#/components/schemas/BackupType'
                restore_file:
                  type: string
                  format: binary
              required:
                - action
      security:
        - BearerAuth: []
      responses:
        200:
          description: Action completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      summary: Get all users
      operationId: getUsers
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create, update, or delete user
      operationId: updateUser
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                operation:
                  $ref: '#/components/schemas/OperationType'
                user:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
              required:
                - operation
                - user
      security:
        - BearerAuth: []
      responses:
        200:
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/import:
    get:
      summary: Get import directory list
      operationId: getImportDirectory
      tags:
        - Admin
      parameters:
        - name: directory
          in: query
          schema:
            type: string
        - name: select
          in: query
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryListResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Perform import
      operationId: postImport
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                directory:
                  type: string
                type:
                  $ref: '#/components/schemas/ImportType'
              required:
                - directory
                - type
      security:
        - BearerAuth: []
      responses:
        200:
          description: Import completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResultsResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/import-results:
    get:
      summary: Get import results
      operationId: getImportResults
      tags:
        - Admin
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResultsResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/logs:
    get:
      summary: Get logs with optional filter
      operationId: getLogs
      tags:
        - Admin
      parameters:
        - name: filter
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
      security:
        - BearerAuth: []
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'